[feat] Add disable apply button functionality
