+++ aim/web/ui/src/components/kit/Input/Input.tsx
@@ -103,6 +103,7 @@ function InputWrapper({
       let metadata: IValidationMetadata = { isValid: true, messages: [] };
       if (!_.isEmpty(validationPatterns)) {
         metadata = validatePatterns(validationPatterns, newValue);
+        setIsInputValid(metadata.isValid);
       }
       onChange && onChange(e, newValue, metadata);
     },
@@ -126,7 +127,6 @@ function InputWrapper({
       isRequiredNumberValue,
       isNumberValueFloat,
     });
-    setIsInputValid(true);
     setInputValue(newValue);
     return newValue;
   };
