+++ aim/web/ui/src/modules/BaseExplorer/components/QueryForm/SearchButton/SearchButton.d.ts
@@ -1,4 +1,6 @@
-export interface ISearchButtonProps {
+import { IButtonProps } from 'components/kit/Button/Button.d';
+
+export interface ISearchButtonProps extends IButtonProps {
   isFetching: boolean;
   onSubmit: () => void;
 }
