+++ aim/web/ui/src/modules/BaseExplorer/components/Visualizer/Visualizer.scss
@@ -1,10 +1,11 @@
 .Visualizer {
   width: 100%;
-  height: calc(100vh - 132px);
+  height: 100%;
   display: flex;
+  flex-direction: column;
   .VisualizerContainer {
     width: 100%;
     display: flex;
-    flex-direction: column;
+    height: 100%;
   }
 }
